Update app.js
This change solves the problem where the `for` loop of the login route never looked beyond user 1. It evaluated the first user and either returned the token or a "Username or password is incorrect" message, deeming any further iteration unreachable. By using the proposed `Array.find()` syntax to match a username to a password, we are able to resolve this issue.